refactor: move curated plugin sync into startup_sync
