+++ codex-rs/core/src/plugins/startup_sync_tests.rs
@@ -1,10 +1,17 @@
 use super::*;
+use crate::auth::CodexAuth;
+use crate::config::CONFIG_TOML_FILE;
+use crate::plugins::test_support::TEST_CURATED_PLUGIN_SHA;
+use crate::plugins::test_support::write_curated_plugin_sha;
+use crate::plugins::test_support::write_file;
+use crate::plugins::test_support::write_openai_curated_marketplace;
 use pretty_assertions::assert_eq;
 use std::io::Write;
 use tempfile::tempdir;
 use wiremock::Mock;
 use wiremock::MockServer;
 use wiremock::ResponseTemplate;
+use wiremock::matchers::header;
 use wiremock::matchers::method;
 use wiremock::matchers::path;
 use zip::ZipWriter;
@@ -22,8 +29,8 @@ fn curated_plugins_repo_path_uses_codex_home_tmp_dir() {
 #[test]
 fn read_curated_plugins_sha_reads_trimmed_sha_file() {
     let tmp = tempdir().expect("tempdir");
-    fs::create_dir_all(tmp.path().join(".tmp")).expect("create tmp");
-    fs::write(tmp.path().join(".tmp/plugins.sha"), "abc123\n").expect("write sha");
+    std::fs::create_dir_all(tmp.path().join(".tmp")).expect("create tmp");
+    std::fs::write(tmp.path().join(".tmp/plugins.sha"), "abc123\n").expect("write sha");
 
     assert_eq!(
         read_curated_plugins_sha(tmp.path()).as_deref(),
@@ -83,15 +90,15 @@ async fn sync_openai_plugins_repo_downloads_zipball_and_records_sha() {
 async fn sync_openai_plugins_repo_skips_archive_download_when_sha_matches() {
     let tmp = tempdir().expect("tempdir");
     let repo_path = curated_plugins_repo_path(tmp.path());
-    fs::create_dir_all(repo_path.join(".agents/plugins")).expect("create repo");
-    fs::write(
+    std::fs::create_dir_all(repo_path.join(".agents/plugins")).expect("create repo");
+    std::fs::write(
         repo_path.join(".agents/plugins/marketplace.json"),
         r#"{"name":"openai-curated","plugins":[]}"#,
     )
     .expect("write marketplace");
-    fs::create_dir_all(tmp.path().join(".tmp")).expect("create tmp");
+    std::fs::create_dir_all(tmp.path().join(".tmp")).expect("create tmp");
     let sha = "fedcba9876543210fedcba9876543210fedcba98";
-    fs::write(tmp.path().join(".tmp/plugins.sha"), format!("{sha}\n")).expect("write sha");
+    std::fs::write(tmp.path().join(".tmp/plugins.sha"), format!("{sha}\n")).expect("write sha");
 
     let server = MockServer::start().await;
     Mock::given(method("GET"))
@@ -119,6 +126,76 @@ async fn sync_openai_plugins_repo_skips_archive_download_when_sha_matches() {
 
     assert_eq!(read_curated_plugins_sha(tmp.path()).as_deref(), Some(sha));
     assert!(repo_path.join(".agents/plugins/marketplace.json").is_file());
+}
+
+#[tokio::test]
+async fn startup_remote_plugin_sync_writes_marker_and_reconciles_state() {
+    let tmp = tempdir().expect("tempdir");
+    let curated_root = curated_plugins_repo_path(tmp.path());
+    write_openai_curated_marketplace(&curated_root, &["linear"]);
+    write_curated_plugin_sha(tmp.path());
+    write_file(
+        &tmp.path().join(CONFIG_TOML_FILE),
+        r#"[features]
+plugins = true
+
+[plugins."linear@openai-curated"]
+enabled = false
+"#,
+    );
+
+    let server = MockServer::start().await;
+    Mock::given(method("GET"))
+        .and(path("/backend-api/plugins/list"))
+        .and(header("authorization", "Bearer Access Token"))
+        .and(header("chatgpt-account-id", "account_id"))
+        .respond_with(ResponseTemplate::new(200).set_body_string(
+            r#"[
+  {"id":"1","name":"linear","marketplace_name":"openai-curated","version":"1.0.0","enabled":true}
+]"#,
+        ))
+        .mount(&server)
+        .await;
+
+    let mut config = crate::plugins::test_support::load_plugins_config(tmp.path()).await;
+    config.chatgpt_base_url = format!("{}/backend-api/", server.uri());
+    let manager = Arc::new(PluginsManager::new(tmp.path().to_path_buf()));
+    let auth_manager =
+        AuthManager::from_auth_for_testing(CodexAuth::create_dummy_chatgpt_auth_for_testing());
+
+    start_startup_remote_plugin_sync_once(
+        Arc::clone(&manager),
+        tmp.path().to_path_buf(),
+        config,
+        auth_manager,
+    );
+
+    let marker_path = tmp.path().join(STARTUP_REMOTE_PLUGIN_SYNC_MARKER_FILE);
+    tokio::time::timeout(Duration::from_secs(5), async {
+        loop {
+            if marker_path.is_file() {
+                break;
+            }
+            tokio::time::sleep(Duration::from_millis(10)).await;
+        }
+    })
+    .await
+    .expect("marker should be written");
+
+    assert!(
+        tmp.path()
+            .join(format!(
+                "plugins/cache/openai-curated/linear/{TEST_CURATED_PLUGIN_SHA}"
+            ))
+            .is_dir()
+    );
+    let config =
+        std::fs::read_to_string(tmp.path().join(CONFIG_TOML_FILE)).expect("config should exist");
+    assert!(config.contains(r#"[plugins."linear@openai-curated"]"#));
+    assert!(config.contains("enabled = true"));
+
+    let marker_contents = std::fs::read_to_string(marker_path).expect("marker should be readable");
+    assert_eq!(marker_contents, "ok\n");
 }
 
 fn curated_repo_zipball_bytes(sha: &str) -> Vec<u8> {
